+++ src/main/java/com/hiddless/Controller/TeacherController.java
@@ -3,8 +3,10 @@ package com.hiddless.Controller;
 import com.hiddless.dao.IDaoGenerics;
 import com.hiddless.dao.TeacherDao;
 import com.hiddless.dto.TeacherDto;
+import com.hiddless.log.LogExecutionTime;
 import com.hiddless.utils.SpecialColours;
 
+import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
@@ -21,45 +23,68 @@ public class TeacherController implements IDaoGenerics<TeacherDto> {
     /// Create
     @Override
     public Optional<TeacherDto> create(TeacherDto teacherDto){
-        Optional<TeacherDto> createdTeacher = teacherDao.create(teacherDto);
-        if (createdTeacher == null) {
-            System.out.println(SpecialColours.RED + "Failed to create a teacher.Please Try again."+ SpecialColours.RESET);
+        if (teacherDto == null || teacherDao.findById(teacherDto.id()).isPresent()) {
+            System.out.println(SpecialColours.RED + "Invalid Teacher Id" + SpecialColours.RESET);
+            return Optional.empty();
         }
-        return createdTeacher;
+        return teacherDao.create(teacherDto);
     }
 
     /// Find By name
     @Override
     public Optional<TeacherDto> findByName(String name) {
+        if (name == null || name.trim().isEmpty()) {
+            throw new IllegalArgumentException("Invalid name");
+        }
         return teacherDao.findByName(name);
     }
 
+    /// Find by id
     @Override
     public Optional<TeacherDto> findById(int id) {
-        return null;
+        if (id <= 0) {
+            throw new IllegalArgumentException("Invalid Id");
+        }
+        return teacherDao.findById(id);
     }
 
     /// List
     @Override
+    @LogExecutionTime
     public List<TeacherDto> list() {
+        List<TeacherDto> teacherDtoList = Optional.of(teacherDao.list()).orElse(Collections.emptyList());
+        if (teacherDtoList.isEmpty()) {
+            System.out.println(SpecialColours.RED + "There is no Teacher in list" + SpecialColours.RESET);
+        }
         return teacherDao.list();
     }
 
     /// Update
     @Override
     public Optional<TeacherDto> update(int id, TeacherDto teacherDto) {
+        if (id <= 0 || teacherDto == null) {
+            throw new IllegalArgumentException("Please write an valid Teacher");
+        }
         return teacherDao.update(id, teacherDto);
     }
 
     /// Delete
     @Override
     public Optional<TeacherDto> delete(int id) {
+        if (id <= 0) {
+            throw new IllegalArgumentException("Please write an valid Teacher ID");
+        }
         return teacherDao.delete(id);
     }
 
     /// Chooise(Switch case)
     @Override
+    @LogExecutionTime
     public void chooise(){
         teacherDao.chooise();
+    }
+
+    public static void main(String[] args) {
+
     }
 }
